new change

+++ database/migrations/2023_04_16_112746_create_abonnements_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             //foreignKey
             $table->unsignedInteger('code_categorie');
             $table->unsignedInteger('num_agence');
-            $table->unsignedInteger('rythme_facturation_id');
+            $table->unsignedBigInteger('rythme_facturation_id');
             $table->string('num_abonne');
             $table->string('num_point_livraison');
             $table->timestamps();
